+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `--alternate-assets` option to the CLI ([#46](https://github.com/stac-utils/stac-asset/pull/46))
 - Content-type checking ([#46](https://github.com/stac-utils/stac-asset/pull/46))
 - `Client.from_config` and `Client.close` ([#46](https://github.com/stac-utils/stac-asset/pull/46))
+- `Collection` download ([#50](https://github.com/stac-utils/stac-asset/pull/50))
 
 ### Changed